requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Scheduler
apscheduler==3.10.4

# WeChat crypto
pycryptodome==3.21.0
lxml==5.3.0

# HTTP client
httpx==0.28.1
aiofiles==24.1.0

# Markdown
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2

# Utils
python-dotenv==1.0.1
openpyxl==3.1.5
Pillow==11.1.0
